--- bridge-firmwareOOP/NunchuckFunctions.h
@@ -1,135 +1,0 @@
-/*
- * Nunchuck functions  -- Talk to a Wii Nunchuck
- *
- * Original by:
- * 2007 Tod E. Kurt, http://todbot.com/blog/
- *
- * Adapted by:
- * 2021 Lonn,  github.com/lonnwolf031
- *
- */
-
-#include <Arduino.h>
-#include <Wire.h>
-
-class NunchuckFunctions {
-
-public:
-
-  uint8_t nunchuck_buf[6];   // array to store nunchuck data,
-
-  // initialize the I2C system, join the I2C bus,
-  // and tell the nunchuck we're talking to it
-  void nunchuck_init(void);
-
-  // Send a request for data to the nunchuck
-  void nunchuck_send_request(void);
-
-
-  // Encode data to format that most wiimote drivers except
-  // only needed if you use one of the regular wiimote drivers
-  char nunchuk_decode_byte (char x);
-  // Receive data back from the nunchuck,
-  // returns 1 on successful read. returns 0 on failure
-  int nunchuck_get_data(void) const {if (cnt >= 5) {return 1;   // success
-  }
-  return 0; //failure}
-  {
-      int cnt=0;
-      Wire.requestFrom (0x52, 6);// request data from nunchuck
-      while (Wire.available ()) {
-          // receive byte as an integer
-          nunchuck_buf[cnt] = nunchuk_decode_byte(Wire.read());
-          cnt++;
-      }
-      nunchuck_send_request();  // send request for next data payload
-      // If we recieved the 6 bytes, then go print them
-      if (cnt >= 5) {
-          return 1;   // success
-      }
-      return 0; //failure
-  }
-
-  void selectNunchuckChannel(int channel) ;
-
-
-  void nunchuck_print_data()
-  {
-      static int i=0;
-      int joy_x_axis = nunchuck_buf[0];
-      int joy_y_axis = nunchuck_buf[1];
-      int accel_x_axis = nunchuck_buf[2]; // * 2 * 2;
-      int accel_y_axis = nunchuck_buf[3]; // * 2 * 2;
-      int accel_z_axis = nunchuck_buf[4]; // * 2 * 2;
-
-      int z_button = 0;
-      int c_button = 0;
-
-      // byte nunchuck_buf[5] contains bits for z and c buttons
-      // it also contains the least significant bits for the accelerometer data
-      // so we have to check each bit of byte outbuf[5]
-      if ((nunchuck_buf[5] >> 0) & 1)
-          z_button = 1;
-      if ((nunchuck_buf[5] >> 1) & 1)
-          c_button = 1;
-
-      if ((nunchuck_buf[5] >> 2) & 1)
-          accel_x_axis += 2;
-      if ((nunchuck_buf[5] >> 3) & 1)
-          accel_x_axis += 1;
-
-      if ((nunchuck_buf[5] >> 4) & 1)
-          accel_y_axis += 2;
-      if ((nunchuck_buf[5] >> 5) & 1)
-          accel_y_axis += 1;
-
-      if ((nunchuck_buf[5] >> 6) & 1)
-          accel_z_axis += 2;
-      if ((nunchuck_buf[5] >> 7) & 1)
-          accel_z_axis += 1;
-      i++;
-  }
-
-  // returns zbutton state: 1=pressed, 0=notpressed
-  int nunchuck_zbutton()
-  {
-      return ((nunchuck_buf[5] >> 0) & 1) ? 0 : 1;  // voodoo
-  }
-
-  // returns zbutton state: 1=pressed, 0=notpressed
-  int nunchuck_cbutton()
-  {
-      return ((nunchuck_buf[5] >> 1) & 1) ? 0 : 1;  // voodoo
-  }
-
-  // returns value of x-axis joystick
-  int nunchuck_joyx()
-  {
-      return nunchuck_buf[0];
-  }
-
-  // returns value of y-axis joystick
-  int nunchuck_joyy()
-  {
-      return nunchuck_buf[1];
-  }
-
-  // returns value of x-axis accelerometer
-  int nunchuck_accelx()
-  {
-      return nunchuck_buf[2];   // FIXME: this leaves out 2-bits of the data
-  }
-
-  // returns value of y-axis accelerometer
-  int nunchuck_accely()
-  {
-      return nunchuck_buf[3];   // FIXME: this leaves out 2-bits of the data
-  }
-
-  // returns value of z-axis accelerometer
-  int nunchuck_accelz()
-  {
-      return nunchuck_buf[4];   // FIXME: this leaves out 2-bits of the data
-  }
-
-};
