+++ bridge-firmwareOOP/TODOversionOOP.txt
@@ -6,11 +6,10 @@ TODO:
 
 readInput in operate();
 
-Why int, what does this do? (name suggests boolean, because 'released' can be either true or false):
-  int LeftReleased, RightReleased;
+W
 
 Rename elements of m_PWMpins:
-  e.g. "PWM1" doesn't say anything about what is connected to this pin. Couldn't derive this from original code yet.
+joy (nunchuck?) xyz
 
 
 
@@ -23,11 +22,6 @@ void init() {
 
   what are pin # 20 and 21 for?
 
-// analog data == 0 or 1?
+switch leftmode, rightmode
 
-
-// digital read of analog data without map?
-  c_button[0] = !digitalRead(m_analogpins["c_button1"]);
-  z_button[0] = !digitalRead(m_analogpins["z_button1"]);
-  c_button[1] = !digitalRead(m_analogpins["c_button2"]);
-  z_button[1] = !digitalRead(m_analogpins["z_button2"]);
+Rezip, compile and install NunchuckFunctions library
