+++ bridge-firmwareOOP/libraries/NunchuckFunctions/library.properties
@@ -1,0 +1,9 @@
+name=NunchuckFunctions
+version=0.0.1
+author=lonnwolf031
+maintainer=github.com/lonnwolf031
+sentence=Arduino library for controlling Nunchuck
+paragraph=Arduino library for controlling single-wire-based LED pixels and strip.
+category=Device Control
+url=https://github.com/lonnwolf031/splitcontroller/NunchuckFunctions
+architectures=*
